refactor: make jobs/adAssetStore importable without CLI
