+++ packages/@openupm/ads/package.json
@@ -37,7 +37,7 @@
     "test:watch": "cross-env NODE_ENV=test NODE_OPTIONS=--experimental-vm-modules jest --watch"
   },
   "bin": {
-    "adAssetStore": "build/jobs/adAssetStore.js"
+    "adAssetStore": "build/jobs/adAssetStoreCLI.js"
   },
   "dependencies": {
     "@commander-js/extra-typings": "^11.1.0",
